Remove faux 'yaml' package
I never liked the idea of having a fake package mimicking behavior
of https://www.npmjs.com/package/yaml

Besides, the need for parsing/serializing into YAML is rather niche.

With the help of external JS bundlers/transpilers it is possible to
run NPM "YAML" package without any "hacks" on the server-side.

I added an example showing how to do it.

For this to work, we should be able to disable server-side Babel
transpilation, otherwise the function will time out.